refactor: divide cli.go into files by each command
